Refactor GitHub artifact downloads out of `dev-cmd/pr-pull`
I plan to use these in `test-bot` to download built bottles from
previous CI runs.
