+++ Library/Homebrew/download_strategy.rb
@@ -633,6 +633,40 @@ class CurlGitHubPackagesDownloadStrategy < CurlDownloadStrategy
   end
 end
 
+# Strategy for downloading an artifact from GitHub Actions.
+#
+# @api private
+class GitHubArtifactDownloadStrategy < AbstractFileDownloadStrategy
+  def fetch(timeout: nil)
+    ohai "Downloading #{url}"
+    if cached_location.exist?
+      puts "Already downloaded: #{cached_location}"
+    else
+      begin
+        curl "--location", "--create-dirs", "--output", temporary_path, url,
+             *meta.fetch(:curl_args, []),
+             secrets: meta.fetch(:secrets, []),
+             timeout: timeout
+      rescue ErrorDuringExecution
+        raise CurlDownloadStrategyError, url
+      end
+      ignore_interrupts do
+        cached_location.dirname.mkpath
+        temporary_path.rename(cached_location)
+        symlink_location.dirname.mkpath
+      end
+    end
+    FileUtils.ln_s cached_location.relative_path_from(symlink_location.dirname), symlink_location, force: true
+  end
+
+  private
+
+  sig { returns(String) }
+  def resolved_basename
+    "artifact.zip"
+  end
+end
+
 # Strategy for downloading a file from an Apache Mirror URL.
 #
 # @api public
