+++ Library/Homebrew/utils/github.rb
@@ -5,6 +5,7 @@ require "uri"
 require "utils/github/actions"
 require "utils/github/api"
 
+require "download_strategy"
 require "system_command"
 
 # Wrapper functions for the GitHub API.
@@ -369,6 +370,28 @@ module GitHub
     end
 
     artifact.last["archive_download_url"]
+  end
+
+  def self.download_artifact(url, artifact_id, dir = Pathname.pwd)
+    odie "Credentials must be set to access the Artifacts API" if API.credentials_type == :none
+
+    token = API.credentials
+    curl_args = ["--header", "Authorization: token #{token}"]
+
+    # Download the artifact as a zip file and unpack it into `dir`. This is
+    # preferred over system `curl` and `tar` as this leverages the Homebrew
+    # cache to avoid repeated downloads of (possibly large) bottles.
+    FileUtils.chdir dir do
+      downloader = GitHubArtifactDownloadStrategy.new(
+        url,
+        "artifact",
+        artifact_id,
+        curl_args: curl_args,
+        secrets:   [token],
+      )
+      downloader.fetch
+      downloader.stage
+    end
   end
 
   def self.public_member_usernames(org, per_page: 100)
